requirements-production.txt
# Production requirements for Adakings Backend API
# Includes base requirements plus production-specific packages

# Include base requirements
-r requirements.txt

# Production Web Server
gunicorn==21.2.0
whitenoise==6.6.0

# Production Database
psycopg2-binary==2.9.9

# Caching
redis==5.0.1
django-redis==5.4.0

# Monitoring & Logging
sentry-sdk[django]==1.38.0

# Security
django-environ==0.11.2

# Performance
django-extensions==3.2.3

# Health Checks
django-health-check==3.17.0

# Static Files Compression
django-compressor==4.4

# API Documentation (Production optimized)
django-cors-headers>=4.3.1

# Rate Limiting
django-ratelimit==4.1.0

# Environment Management
python-decouple==3.8
